+++ arcane/src/arcane/hdf5/Hdf5ReaderWriter.cc
@@ -1,11 +1,11 @@
 ﻿// -*- tab-width: 2; indent-tabs-mode: nil; coding: utf-8-with-signature -*-
 //-----------------------------------------------------------------------------
-// Copyright 2000-2022 CEA (www.cea.fr) IFPEN (www.ifpenergiesnouvelles.com)
+// Copyright 2000-2023 CEA (www.cea.fr) IFPEN (www.ifpenergiesnouvelles.com)
 // See the top-level COPYRIGHT file for details.
 // SPDX-License-Identifier: Apache-2.0
 //-----------------------------------------------------------------------------
 /*---------------------------------------------------------------------------*/
-/* Hdf5ReaderWriter.cc                                         (C) 2000-2022 */
+/* Hdf5ReaderWriter.cc                                         (C) 2000-2023 */
 /*                                                                           */
 /* Lecture/Ecriture au format HDF5.                                          */
 /*---------------------------------------------------------------------------*/
@@ -50,9 +50,9 @@
 #include "arcane/ISerializeMessageList.h"
 #include "arcane/SerializeMessage.h"
 
-#include "arcane/std/Hdf5ReaderWriter.h"
+#include "arcane/hdf5/Hdf5ReaderWriter.h"
 
-#include "arcane/std/Hdf5ReaderWriter_axl.h"
+#include "arcane/hdf5/Hdf5ReaderWriter_axl.h"
 
 #include <array>
 //#define ARCANE_TEST_HDF5MPI
