+++ arcane/src/arcane/hdf5/srcs.cmake
@@ -3,4 +3,10 @@ set(ARCANE_SOURCES
   Hdf5Utils.h
   Hdf5VariableInfoBase.cc
   Hdf5VariableInfoBase.h
+  Hdf5ReaderWriter.cc
+  Hdf5ReaderWriter.h
+)
+
+set(AXL_FILES
+  Hdf5ReaderWriter
 )
