+++ arcane/src/arcane/std/srcs.cmake
@@ -104,7 +104,6 @@ set(ARCANE_SOURCES
 )
 
 set(AXL_FILES
-  Hdf5ReaderWriter
   Hdf5VariableReader
   Hdf5VariableWriter
   Hdf5MpiReaderWriter
